+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changes
 
+* Enable root `Eventum` module to access `streams` and `events` APIs. ([@enriikke][])
+
 ### Bug Fixes